IMPALA-10640: Support reading Parquet Bloom filters - most common types
This change adds read support for Parquet Bloom filters for types that
can reasonably be supported in Impala. Other types, such as CHAR(N),
would be very difficult to support because the length may be different
in Parquet and in Impala which results in truncation or padding, and
that changes the hash which makes using the Bloom filter impossible.
Write support will be added in a later change.
The supported Parquet type - Impala type pairs are the following:

 ---------------------------------------
|Parquet type |  Impala type            |
|---------------------------------------|
|INT32        |  TINYINT, SMALLINT, INT |
|INT64        |  BIGINT                 |
|FLOAT        |  FLOAT                  |
|DOUBLE       |  DOUBLE                 |
|BYTE_ARRAY   |  STRING                 |
 ---------------------------------------

The following types are not supported for the given reasons:

 ----------------------------------------------------------------
|Impala type |  Problem                                          |
|----------------------------------------------------------------|
|VARCHAR(N)  | truncation can change hash                        |
|CHAR(N)     | padding / truncation can change hash              |
|DECIMAL     | multiple encodings supported                      |
|TIMESTAMP   | multiple encodings supported, timezone conversion |
|DATE        | not considered yet                                |
 ----------------------------------------------------------------

Support may be added for these types later, see IMPALA-10641.

If a Bloom filter is available for a column that is fully dictionary
encoded, the Bloom filter is not used as the dictionary can give exact
results in filtering.

Testing:
  - Added tests/query_test/test_parquet_bloom_filter.py that tests
    whether Parquet Bloom filtering works for the supported types and
    that we do not incorrectly discard row groups for the unsupported
    type VARCHAR. The Parquet file used in the test was generated with
    an external tool.
  - Added unit tests for ParquetBloomFilter in file
    be/src/util/parquet-bloom-filter-test.cc
  - A minor, unrelated change was done in
    be/src/util/bloom-filter-test.cc: the MakeRandom() function had
    return type uint64_t, the documentation claimed it returned a 64 bit
    random number, but the actual number of random bits is 32, which is
    what is intended in the tests. The return type and documentation
    have been corrected to use 32 bits.

Reviewed-on: http://gerrit.cloudera.org:8080/17026
Tested-by: Csaba Ringhofer <csringhofer@cloudera.com>

+++ be/src/util/bloom-filter-test.cc
@@ -41,9 +41,9 @@ using kudu::rpc::RpcController;
 
 namespace bloom_filter_test_util {
 
-// Make a random uint64_t, avoiding the absent high bit and the low-entropy low bits
+// Make a random uint32_t, avoiding the absent high bit and the low-entropy low bits
 // produced by rand().
-uint64_t MakeRand() {
+uint32_t MakeRand() {
   uint32_t result = (rand() >> 8) & 0xffff;
   result <<= 16;
   result |= (rand() >> 8) & 0xffff;
@@ -278,7 +278,7 @@ TEST_F(BloomFilterTest, Find) {
   for (int i = 13; i < 17; ++i) {
     BloomFilter* bf = CreateBloomFilter(i);
     for (int k = 0; k < (1 << 15); ++k) {
-      const uint64_t to_insert = MakeRand();
+      const uint32_t to_insert = MakeRand();
       BfInsert(*bf, to_insert);
       EXPECT_TRUE(BfFind(*bf, to_insert));
     }
